requirements.txt
asgiref==3.6.0
Django==4.1.5
django-taggit==3.1.0
djangorestframework==3.14.0
mysql-connector==2.2.9
mysqlclient==2.1.1
Pillow==9.4.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
